proper update for llvm_asm
I'd like to thank the new contributor for the inspiration. This patch
is a bit more extensive, in that it keeps the older code for people
who may not be updating their nightly toolchain every day. Myself
primarily, I tend to be a little lazy with compiler updates.

There maybe some temporary breakage. If you use the `nightly-1.46.0`
builds which were created before `2020-06-08`.